SAKAI-2193 WARNING. SEE FULL COMMIT MESSAGE
WIP. This is attempting to get reoccurring evnets to be generated.
As it stands tomcat will not start up. Spring can not find the
RRule class. This can be resolved by manaually copying the google.jar
to /opt/tomcat/shared/lib/, this causes the Pair class to not be
found (same resolution as RRule). Then Tomcat will start. Than
when going to the attendance tool this will cause a loader constraint
violation. The basepage can not be constructed.