SAKAI-2199 Add getAttendanceGrades to Logic
This adds the method getAttendanceGrades to the AttendanceLogic. This
method returns a HashMap of the AttendanceGrades for the current
AttendanceSite. If no AGs are present, it will generate AGs for all
current site members. If AGs are present, it will check to see if
all currentUsers are in the AGs and if any are missing it will
generate them.