SAKAI-2366 Improved null list handling
All lists are now checked to make sure not to return null, they will
return empty if null. The DAO will return null if there is a
DataAccessException (i.e. DB dies), or empty if there are no results.
The logic thus will take the null list and return an empty.
Also, the getAttendanceEvents() function was removed...it is
TOO POWERFUL (it would return all attendanceEvents)!