SAKAI-3129 Improve Attendance Performance
Attendance performance was imporoved by now keeping track of statistics
in a table (instead of previously calculating them on the fly).

Stats table which keeps track of the statistics for an individual
attendance item or user (per site). These statistics are the total
number of each Status.

Streamlined the database operations to only save records when necessary
, that is when they actually have content to save.

Improved the StudentOverview page to use one query to the DB.

Minor optimizations.