SAKAI-3132 Add auto grading to Attendance
This is a new feature that allows a user to create rules that
will auto grade a student's attendance based on the student's
statuses. Grading can be done based on students starting with
either zero points and adding points or starting with total
points and losing points. Additionally, the instructor has the
ability to override the auto calculated grade.

Out of scope for this ticket is a grade log, validation for
grading rules, and a way to "test" the created grading rules.

Following messages are from commits squashed into this commit:
Ensured that certain variables within the AttendanceSite where null
safe (all the variables which have been added after Phase 1).

Fixed bugs with feedback messages, and NPEs. Changed the wording
for the grading rules form to be in a dropdown box with a better
explanation. Removed override checkbox and added total to the
student view grade form.

Seperated the grading rules form as clearly as possible. Added a "grade
all" button to show when rule changes are made. Fixed styling and word
issues.

Fix the send to gradebook functionality with regards to the auto
grade functionality.

The logic functionality which handles updating AttendanceGrades will
now send to the gradebook as necessary.

The sendToGradebook functionality now takes an AttendanceGrade item
which should help improve performance, less calls to the DB.