SAKAI-3158 Add Attendance Stat Calc Job
Add a job which will calculate the statistics for AttendanceStats
objects (ItemStats and UserStats).

This job must be run at least once after SAKAI-3129 has been deployed,
as 3129 adds Stats calculations but only for new record saves. This
job will compute the stats for all AttendanceSites in batches of 5.

When an AttendanceSite is calculating, the site enters read only mode.
No records may be modified and no Items may not be deleted. This was
done to prevent incorrect calculations (i.e. records are loaded in
and then changed elsewhere).

When AttendanceSites are marked in sync, the syncTime is also
marked. This will allow for that site to no longer be picked up by
the getAttendanceSiteBatch method as the syncTime will be set to the
current syncTime of said job's run. So, if an error occurs before the
calculations for a site may finish, the Site will not be picked up
in the next batch.

The calculation only occurs for sites which have Users present within
them; if the user list is empty, than either the site has no users
(in which case, why calculate stats) or the site no longer exists (so
why calculate stats).

Records for users which are no longer in the site are ignored.

If debug is on for the job messages will be printed informing the user
of the above issuesi (no users / record of user not present).

If an admin wants to run the job for only a subset of sites, they can
set the syncTime for all the other sites to some point in the future.
The job only selects AttendanceSites whose syncTime is in the past or
null.