SAKAI-3167 Add/Edit Item now in Modal + Refactoring
Add the ability to add/edit/delete an Attendance Item within a modal
window. As such removed the Add/Edit items page.

Refactor panels to be in a better location.
Refactor some variable names to be more internally consistent.

SAKAI-3167 Feedback in modal, and other changes

Changed the delete item from a icon to a text link.
Added some padding to the Wicket Date picker icon.
Updated the text on the main page.

Fixed a bug when editing an item on the same page you are on.

Improve the modal window to now support feedback within the modal.

Also move the Add Another checkbox to a button.